22.1.Conditional Field Setup
So far, we've been talking about granting or denying access to something entirely. In CheeseListing, the most complex case was when we used a voter to deny access to the PUT operation unless you are the owner of this CheeseListing or an admin user.

But there are several other ways that you might need to customize access to your API. For example, what if we have a field that we only want readable, or maybe writable by certain types of users? A great example of this is in User: the  field. Right now, the $roles field is not part of our API at all: nobody can change this field via the API.

That certainly make sense... for most API users. But what if we create an admin section and we do need this field to be editable by admin users? How can we do that?

We'll get there. For now, let's add this to the API for all users by adding @groups("user:write"). This creates a huge security hole... so we'll come back to this in a few minutes and make sure that only admin users can write to this field.

Next, sometimes, based on who is logged in, you might need to show additional fields or hide some fields. The same is true when creating or updating a resource: an admin user might have access to write a field that normal users can't.

Let's start getting this all set up!